# swagger/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marcoo BFF",
    "description": "Marcoo BFF",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Local server",
      "variables": {
        "port": {
          "default": "8082"
        }
      }
    },
    {
      "url": "https://marcoo-bff.herokuapp.com/",
      "description": "(Internal) Production server"
    }
  ],
  "tags": [
    {
      "name": "API",
      "description": "Microservice endpoints"
    },
    {
      "name": "Health check",
      "description": "Service health check"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Check service health",
        "tags": [
          "Health check"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthCheckResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/supermarkets/{cep}/products": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                },
                "examples": {
                  "example-product": {
                    "$ref": "#/components/examples/ProductExample"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "502": {
            "$ref": "#/components/responses/ServiceError"
          }
        },
        "operationId": "get-products",
        "description": "Obtém produtos disponíveis",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "name",
            "description": "Busca pelo nome do produto"
          }
        ]
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cep",
          "in": "path",
          "required": true,
          "description": "Número do CEP"
        }
      ]
    },
    "/v1/supermarkets/{cep}/products/categories/{type}/{category}": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                },
                "examples": {
                  "example-product": {
                    "$ref": "#/components/examples/ProductExample"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "502": {
            "$ref": "#/components/responses/ServiceError"
          }
        },
        "operationId": "get-products-by-categories",
        "description": "Obtém os produtos por categoria",
        "parameters": []
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cep",
          "in": "path",
          "required": true,
          "description": "Número do CEP"
        },
        {
          "schema": {
            "type": "string",
            "enum": [
              "first",
              "second",
              "third"
            ]
          },
          "name": "type",
          "in": "path",
          "required": true,
          "description": "Tipo da categoria"
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "category",
          "in": "path",
          "required": true,
          "description": "Id da categoria"
        }
      ]
    },
    "/v1/supermarkets/{cep}/products/checkout": {
      "post": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "502": {
            "$ref": "#/components/responses/ServiceError"
          }
        },
        "operationId": "get-products-checkout",
        "description": "Obtém a lista de comparação de preços no checkout",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "",
                "type": "object",
                "properties": {
                  "data": {
                    "type": "array",
                    "uniqueItems": true,
                    "minItems": 1,
                    "items": {
                      "required": [
                        "idProduct",
                        "amount"
                      ],
                      "properties": {
                        "idProduct": {
                          "type": "string",
                          "minLength": 1
                        },
                        "amount": {
                          "type": "number"
                        }
                      }
                    }
                  }
                },
                "required": [
                  "data"
                ],
                "x-examples": {
                  "example-1": {
                    "data": [
                      {
                        "idProduct": "asdfsdfsdf",
                        "amount": 2
                      },
                      {
                        "idProduct": "asdfsdfsdf",
                        "amount": 1
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cep",
          "in": "path",
          "required": true,
          "description": "Número do CEP"
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "HealthCheckResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Application running at: 8082"
          },
          "healthy": {
            "type": "boolean",
            "example": true
          },
          "services": {
            "type": "object",
            "additionalProperties": {
              "description": "A map of services, where the keys correspond to service names, and the value to it's health",
              "type": "boolean",
              "example": true
            }
          }
        }
      },
      "ProductResponse": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "data": [
              {
                "_id": "609330b0642b8120248dd5fc",
                "name": "Cerveja BUDWEISER Lata 350ml",
                "brand": "Budweiser",
                "categories": {
                  "first": "60932a930710622024cad153",
                  "second": "60932b250710622024cad1a1",
                  "third": "60932bdb0710622024cad36a"
                },
                "image": "https://i.ibb.co/3sDQSbD/bud-latinha.jpg",
                "material": "Lata",
                "volume": "350 ml",
                "amount": "1",
                "prices": [
                  {
                    "_id": "609589da0710622024cad3a3",
                    "id_supermarket": "60848b9c504e7f1bb1feb73d",
                    "name": "São Vicente",
                    "price": 2.65
                  },
                  {
                    "_id": "609589da0710622024cad3a1",
                    "id_supermarket": "608481c0504e7f1bb1feb73a",
                    "name": "Carrefour",
                    "price": 2.9
                  },
                  {
                    "_id": "609589da0710622024cad3a4",
                    "id_supermarket": "60848bc7504e7f1bb1feb73e",
                    "name": "Dia%",
                    "price": 2.91
                  },
                  {
                    "_id": "609589da0710622024cad3a5",
                    "id_supermarket": "60848c10504e7f1bb1feb73f",
                    "name": "Makro Atacadista",
                    "price": 2.91
                  },
                  {
                    "_id": "609589da0710622024cad3a0",
                    "id_supermarket": "60845d5468c33834de710e19",
                    "name": "Hipermercado Extra",
                    "price": 2.94
                  }
                ]
              }
            ]
          }
        },
        "properties": {
          "data": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 0,
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "minLength": 1
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "brand": {
                  "type": "string",
                  "minLength": 1
                },
                "categories": {
                  "type": "object",
                  "properties": {
                    "first": {
                      "type": "string",
                      "minLength": 1
                    },
                    "second": {
                      "type": "string",
                      "minLength": 1
                    },
                    "third": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "first",
                    "second",
                    "third"
                  ]
                },
                "image": {
                  "type": "string",
                  "minLength": 1
                },
                "material": {
                  "type": "string",
                  "minLength": 1
                },
                "volume": {
                  "type": "string",
                  "minLength": 1
                },
                "amount": {
                  "type": "string",
                  "minLength": 1
                },
                "prices": {
                  "type": "array",
                  "uniqueItems": true,
                  "minItems": 0,
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "id_supermarket": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "price": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "_id",
                      "id_supermarket",
                      "name",
                      "price"
                    ]
                  }
                }
              },
              "required": [
                "_id",
                "name",
                "brand",
                "image",
                "material",
                "volume",
                "amount"
              ]
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "CheckoutResponse": {
        "description": "",
        "type": "object",
        "properties": {
          "supermarkets": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "required": [
                "name",
                "total"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "total": {
                  "type": "number"
                },
                "products": {
                  "type": "array",
                  "uniqueItems": true,
                  "minItems": 1,
                  "items": {
                    "required": [
                      "idProduct",
                      "name",
                      "amount",
                      "price",
                      "total"
                    ],
                    "properties": {
                      "idProduct": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "amount": {
                        "type": "number"
                      },
                      "price": {
                        "type": "number"
                      },
                      "total": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "required": [
          "supermarkets"
        ],
        "x-examples": {
          "example-1": {
            "supermarkets": [
              {
                "name": "Marcoo",
                "total": 74.61,
                "products": [
                  {
                    "idProduct": "609550324e3ef6461c0b0f8c",
                    "name": "Cerveja HEINEKEN Garrafa 330ml",
                    "amount": 1,
                    "price": 3.56,
                    "total": 3.56
                  },
                  {
                    "idProduct": "609582bd0710622024cad380",
                    "name": "Cerveja STELLA ARTOIS Long Neck 275ML",
                    "amount": 1,
                    "price": 1.69,
                    "total": 1.69
                  },
                  {
                    "idProduct": "609583600710622024cad385",
                    "name": "Cerveja COLORADO Ribeirão Lager 600 ML Garrafa",
                    "amount": 1,
                    "price": 8.32,
                    "total": 8.32
                  },
                  {
                    "idProduct": "609583600710622024cad392",
                    "name": "Refrigerante ANTARCTICA Guaraná Zero Lata 350ML 12 unidades",
                    "amount": 4,
                    "price": 15.26,
                    "total": 61.04
                  }
                ]
              },
              {
                "name": "São Vicente",
                "total": 77.05,
                "products": [
                  {
                    "idProduct": "609582bd0710622024cad380",
                    "name": "Cerveja STELLA ARTOIS Long Neck 275ML",
                    "amount": 1,
                    "price": 1.81,
                    "total": 1.81
                  },
                  {
                    "idProduct": "609550324e3ef6461c0b0f8c",
                    "name": "Cerveja HEINEKEN Garrafa 330ml",
                    "amount": 1,
                    "price": 4.53,
                    "total": 4.53
                  },
                  {
                    "idProduct": "609583600710622024cad385",
                    "name": "Cerveja COLORADO Ribeirão Lager 600 ML Garrafa",
                    "amount": 1,
                    "price": 9.67,
                    "total": 9.67
                  },
                  {
                    "idProduct": "609583600710622024cad392",
                    "name": "Refrigerante ANTARCTICA Guaraná Zero Lata 350ML 12 unidades",
                    "amount": 4,
                    "price": 15.26,
                    "total": 61.04
                  }
                ]
              },
              {
                "name": "Hipermercado Extra",
                "total": 82.42,
                "products": [
                  {
                    "idProduct": "609582bd0710622024cad380",
                    "name": "Cerveja STELLA ARTOIS Long Neck 275ML",
                    "amount": 1,
                    "price": 1.79,
                    "total": 1.79
                  },
                  {
                    "idProduct": "609550324e3ef6461c0b0f8c",
                    "name": "Cerveja HEINEKEN Garrafa 330ml",
                    "amount": 1,
                    "price": 4.01,
                    "total": 4.01
                  },
                  {
                    "idProduct": "609583600710622024cad385",
                    "name": "Cerveja COLORADO Ribeirão Lager 600 ML Garrafa",
                    "amount": 1,
                    "price": 9.58,
                    "total": 9.58
                  },
                  {
                    "idProduct": "609583600710622024cad392",
                    "name": "Refrigerante ANTARCTICA Guaraná Zero Lata 350ML 12 unidades",
                    "amount": 4,
                    "price": 16.76,
                    "total": 67.04
                  }
                ]
              },
              {
                "name": "Makro Atacadista",
                "total": 82.73,
                "products": [
                  {
                    "idProduct": "609582bd0710622024cad380",
                    "name": "Cerveja STELLA ARTOIS Long Neck 275ML",
                    "amount": 1,
                    "price": 1.85,
                    "total": 1.85
                  },
                  {
                    "idProduct": "609550324e3ef6461c0b0f8c",
                    "name": "Cerveja HEINEKEN Garrafa 330ml",
                    "amount": 1,
                    "price": 3.56,
                    "total": 3.56
                  },
                  {
                    "idProduct": "609583600710622024cad392",
                    "name": "Refrigerante ANTARCTICA Guaraná Zero Lata 350ML 12 unidades",
                    "amount": 4,
                    "price": 17.25,
                    "total": 69
                  },
                  {
                    "idProduct": "609583600710622024cad385",
                    "name": "Cerveja COLORADO Ribeirão Lager 600 ML Garrafa",
                    "amount": 1,
                    "price": 8.32,
                    "total": 8.32
                  }
                ]
              }
            ]
          }
        }
      }
    },
    "parameters": {},
    "responses": {
      "ForbiddenError": {
        "description": "Não possui permissão de acesso ao recurso",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erro interno no servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceError": {
        "description": "Erro de serviço",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Erro de requisição mal formada",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "examples": {
      "ProductExample": {
        "value": {
          "data": [
            {
              "_id": "609330b0642b8120248dd5fc",
              "name": "Cerveja BUDWEISER Lata 350ml",
              "brand": "Budweiser",
              "categories": {
                "first": "60932a930710622024cad153",
                "second": "60932b250710622024cad1a1",
                "third": "60932bdb0710622024cad36a"
              },
              "image": "https://i.ibb.co/3sDQSbD/bud-latinha.jpg",
              "material": "Lata",
              "volume": "350 ml",
              "amount": "1",
              "prices": [
                {
                  "_id": "609589da0710622024cad3a3",
                  "id_supermarket": "60848b9c504e7f1bb1feb73d",
                  "name": "São Vicente",
                  "price": 2.65
                },
                {
                  "_id": "609589da0710622024cad3a1",
                  "id_supermarket": "608481c0504e7f1bb1feb73a",
                  "name": "Carrefour",
                  "price": 2.9
                },
                {
                  "_id": "609589da0710622024cad3a4",
                  "id_supermarket": "60848bc7504e7f1bb1feb73e",
                  "name": "Dia%",
                  "price": 2.91
                },
                {
                  "_id": "609589da0710622024cad3a5",
                  "id_supermarket": "60848c10504e7f1bb1feb73f",
                  "name": "Makro Atacadista",
                  "price": 2.91
                },
                {
                  "_id": "609589da0710622024cad3a0",
                  "id_supermarket": "60845d5468c33834de710e19",
                  "name": "Hipermercado Extra",
                  "price": 2.94
                }
              ]
            }
          ]
        }
      }
    }
  }
}